requirements.txt
fastapi==0.116.1
uvicorn==0.18.3
google-generativeai==0.8.5
python-dotenv==1.1.1
pdfplumber==0.11.7
python-docx==1.2.0
numpy==2.3.2
psycopg2-binary==2.9.7
supabase==1.0.0
requests==2.32.4
python-multipart==0.0.20
faiss-cpu==1.11.0.post1
pydantic==2.11.7
